Allow _default.php as alternative to _notfound.php
`_default.php` has priority, and provides a more descriptive name when the intent is to define a dynamic route handler that normally handles a lot of requests where `_notfound.php` would feel like a misnomer, implying handling an exceptional condition.

If the "normal" HTTP response code for this handler going to be 200, it should probably be named `_default.php`. Conversely, if the "normal" response code is 404, name it `_notfound.php`.